pyproject.toml
[project]
name = "magical-utility"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "black>=25.1.0",
    "fake-useragent>=2.1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "ruff>=0.11.0",
    "scrapy-fake-useragent>=1.4.4",
    "scrapy>=2.12.0",
    "tldextract>=5.1.3",
    "scrapy-rotating-proxies>=0.6.2",
    "scrapy-playwright>=0.0.43",
    "ragas>=0.2.14",
    "scipy>=1.15.2",
    "arize-phoenix>=8.24.2",
    "python-levenshtein>=0.27.3",
]
